+++ TractorMod/Framework/Config/FertilizerConfig.cs
@@ -1,0 +1,9 @@
+namespace Pathoschild.Stardew.TractorMod.Framework.Config
+{
+    /// <summary>Configuration for the fertilizer attachment.</summary>
+    internal class FertilizerConfig
+    {
+        /// <summary>Whether to apply fertilizer to tilled dirt.</summary>
+        public bool Enable { get; set; } = true;
+    }
+}