+++ TractorMod/Framework/Config/GrassStarterConfig.cs
@@ -1,0 +1,9 @@
+namespace Pathoschild.Stardew.TractorMod.Framework.Config
+{
+    /// <summary>Configuration for the grass starter attachment.</summary>
+    internal class GrassStarterConfig
+    {
+        /// <summary>Whether to plant grass starters in tilled dirt.</summary>
+        public bool Enable { get; set; } = true;
+    }
+}