+++ TractorMod/Framework/Config/SeedBagConfig.cs
@@ -1,0 +1,9 @@
+namespace Pathoschild.Stardew.TractorMod.Framework.Config
+{
+    /// <summary>Configuration for the Seed Bag mod attachment.</summary>
+    internal class SeedBagModConfig
+    {
+        /// <summary>Whether to use the SeedBag.</summary>
+        public bool Enable { get; set; } = true;
+    }
+}