+++ ContentPatcher/release-notes.md
@@ -3,6 +3,8 @@
 # Release notes
 ## 1.1
 * Added `Enabled` field to easily disable individual patches.
+* Fixed `FromFile` not accepting XNB files unless you remove the extension.
+* Fixed error when loading a local map which references PNG tilesheets.
 
 ## 1.0
 * Initial release.